+++ main.go
@@ -17,6 +17,7 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	restclient "k8s.io/client-go/rest"
+	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 	capiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -27,6 +28,7 @@ import (
 	"github.com/giantswarm/azure-admission-controller/pkg/azuremachinepool"
 	"github.com/giantswarm/azure-admission-controller/pkg/azureupdate"
 	"github.com/giantswarm/azure-admission-controller/pkg/cluster"
+	"github.com/giantswarm/azure-admission-controller/pkg/machinepool"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
@@ -64,6 +66,7 @@ func mainError() error {
 				providerv1alpha1.AddToScheme,
 				corev1alpha1.AddToScheme,
 				releasev1alpha1.AddToScheme,
+				expcapzv1alpha3.AddToScheme,
 			},
 			Logger: newLogger,
 
@@ -210,6 +213,30 @@ func mainError() error {
 		}
 	}
 
+	var machinePoolCreateValidator *machinepool.CreateValidator
+	{
+		c := machinepool.CreateValidatorConfig{
+			CtrlClient: ctrlClient,
+			Logger:     newLogger,
+			VMcaps:     vmcaps,
+		}
+		machinePoolCreateValidator, err = machinepool.NewCreateValidator(c)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	var machinePoolUpdateValidator *machinepool.UpdateValidator
+	{
+		c := machinepool.UpdateValidatorConfig{
+			Logger: newLogger,
+		}
+		machinePoolUpdateValidator, err = machinepool.NewUpdateValidator(c)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
 	// Here we register our endpoints.
 	handler := http.NewServeMux()
 	// Mutators.
@@ -224,6 +251,8 @@ func mainError() error {
 	handler.Handle("/validate/azuremachinepool/create", validator.Handler(azureMachinePoolCreateValidator))
 	handler.Handle("/validate/azuremachinepool/update", validator.Handler(azureMachinePoolUpdateValidator))
 	handler.Handle("/validate/cluster/update", validator.Handler(clusterUpdateValidator))
+	handler.Handle("/validate/machinepool/create", validator.Handler(machinePoolCreateValidator))
+	handler.Handle("/validate/machinepool/update", validator.Handler(machinePoolUpdateValidator))
 	handler.HandleFunc("/healthz", healthCheck)
 
 	newLogger.LogCtx(context.Background(), "level", "debug", "message", fmt.Sprintf("Listening on port %s", cfg.Address))